+++ migrations/sqlite/20160825135747_create_followings/up.sql
@@ -1,0 +1,6 @@
+CREATE TABLE followings (
+  user_id INTEGER NOT NULL,
+  post_id INTEGER NOT NULL,
+  email_notifications BOOLEAN NOT NULL DEFAULT 0,
+  PRIMARY KEY (user_id, post_id)
+);