+++ CHANGELOG.md
@@ -7,6 +7,8 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 ### Added
 
+* Added support for composite primary keys.
+
 * Added support for PostgreSQL `NULLS FIRST` and `NULLS LAST` when sorting.
   See http://docs.diesel.rs/diesel/prelude/trait.SortExpressionMethods.html
   for details.